+++ src/device/curtain.ts
@@ -4,6 +4,9 @@ import { SwitchBotPlatform } from '../platform';
 import { debounceTime, skipWhile, take, tap } from 'rxjs/operators';
 import { Service, PlatformAccessory, CharacteristicValue } from 'homebridge';
 import { DeviceURL, device, devicesConfig, serviceData, switchbot, deviceStatusResponse, payload, deviceStatus, ad } from '../settings';
+import { Context } from 'vm';
+import { MqttClient } from 'mqtt';
+import { connectAsync } from 'async-mqtt';
 
 export class Curtain {
   // Services
@@ -24,6 +27,8 @@ export class Curtain {
   slidePosition: deviceStatus['slidePosition'];
   moving: deviceStatus['moving'];
   brightness: deviceStatus['brightness'];
+  setPositionMode?: string | number;
+  Mode!: string;
 
   // BLE Others
   connected?: boolean;
@@ -49,10 +54,15 @@ export class Curtain {
   scanDuration!: number;
   deviceLogging!: string;
   deviceRefreshRate!: number;
+  setCloseMode!: string;
+  setOpenMode!: string;
 
   // Updates
   curtainUpdateInProgress!: boolean;
   doCurtainUpdate!: Subject<void>;
+
+  //MQTT stuff
+  mqttClient: MqttClient | null = null;
 
   constructor(private readonly platform: SwitchBotPlatform, private accessory: PlatformAccessory, public device: device & devicesConfig) {
     // default placeholders
@@ -60,6 +70,7 @@ export class Curtain {
     this.refreshRate(device);
     this.scan(device);
     this.config(device);
+    this.setupMqtt(device);
     this.CurrentPosition = 0;
     this.TargetPosition = 0;
     this.PositionState = this.platform.Characteristic.PositionState.STOPPED;
@@ -77,7 +88,9 @@ export class Curtain {
       .getService(this.platform.Service.AccessoryInformation)!
       .setCharacteristic(this.platform.Characteristic.Manufacturer, 'SwitchBot')
       .setCharacteristic(this.platform.Characteristic.Model, 'W0701600')
-      .setCharacteristic(this.platform.Characteristic.SerialNumber, device.deviceId);
+      .setCharacteristic(this.platform.Characteristic.SerialNumber, device.deviceId)
+      .setCharacteristic(this.platform.Characteristic.FirmwareRevision, this.FirmwareRevision(accessory, device))
+      .getCharacteristic(this.platform.Characteristic.FirmwareRevision).updateValue(this.FirmwareRevision(accessory, device));
 
     // get the WindowCovering service if it exists, otherwise create a new WindowCovering service
     // you can create multiple services for each accessory
@@ -192,6 +205,35 @@ export class Curtain {
         }
         this.curtainUpdateInProgress = false;
       });
+  }
+
+  /*
+   * Publish MQTT message for topics of
+   * 'homebridge-switchbot/curtain/xx:xx:xx:xx:xx:xx'
+   */
+  mqttPublish(topic: string, message: any) {
+    const mac = this.device.deviceId?.toLowerCase().match(/[\s\S]{1,2}/g)?.join(':');
+    const options = this.device.mqttPubOptions || {};
+    this.mqttClient?.publish(`homebridge-switchbot/curtain/${mac}/${topic}`, `${message}`, options);
+    this.debugLog(`Meter: ${this.accessory.displayName} MQTT message: ${topic}/${message} options:${JSON.stringify(options)}`);
+  }
+
+  /*
+   * Setup MQTT hadler if URL is specifed.
+   */
+  async setupMqtt(device: device & devicesConfig): Promise<void> {
+    if (device.mqttURL) {
+      try {
+	this.mqttClient = await connectAsync(device.mqttURL, device.mqttOptions || {});
+	this.debugLog(`Meter: ${this.accessory.displayName} MQTT connection has been established successfully.`)
+	this.mqttClient.on('error', (e: Error) => {
+	  this.errorLog(`Meter: ${this.accessory.displayName} Failed to publish MQTT messages. ${e}`)
+	});
+      } catch (e) {
+	this.mqttClient = null;
+	this.errorLog(`Meter: ${this.accessory.displayName} Failed to establish MQTT connection. ${e}`)
+      }
+    }
   }
 
   /**
@@ -528,12 +570,34 @@ export class Curtain {
         .join(':')
         .toLowerCase();
       this.debugLog(`Curtain: ${this.accessory.displayName} BLE Address: ${this.device.bleMac}`);
+      if (this.TargetPosition > 50) {
+        if (this.device.curtain?.setOpenMode === '1') {
+          this.setPositionMode = 1;
+          this.Mode = 'Silent Mode';
+        } else {
+          this.setPositionMode = 0;
+          this.Mode = 'Performance Mode';
+        }
+      } else {
+        if (this.device.curtain?.setCloseMode === '1') {
+          this.setPositionMode = 1;
+          this.Mode = 'Silent Mode';
+        } else {
+          this.setPositionMode = 0;
+          this.Mode = 'Performance Mode';
+        }
+      }
+      const adjustedMode = this.setPositionMode || null;
+      if (adjustedMode === null) {
+        this.Mode = 'Default Mode';
+      }
+      this.debugLog(`${this.accessory.displayName} Mode: ${this.Mode}`);
       if (switchbot !== false) {
         switchbot
           .discover({ model: 'c', quick: true, id: this.device.bleMac })
           .then((device_list) => {
             this.infoLog(`${this.accessory.displayName} Target Position: ${this.TargetPosition}`);
-            return device_list[0].runToPos(100 - Number(this.TargetPosition));
+            return device_list[0].runToPos(100 - Number(this.TargetPosition), adjustedMode);
           })
           .then(() => {
             this.debugLog(`Curtain: ${this.accessory.displayName} Done.`);
@@ -573,10 +637,24 @@ export class Curtain {
         if (this.TargetPosition !== this.CurrentPosition) {
           this.debugLog(`Pushing ${this.TargetPosition}`);
           const adjustedTargetPosition = 100 - Number(this.TargetPosition);
+          if (this.TargetPosition > 50) {
+            this.setPositionMode = this.device.curtain?.setOpenMode;
+          } else {
+            this.setPositionMode = this.device.curtain?.setCloseMode;
+          }
+          if (this.setPositionMode === '1') {
+            this.Mode = 'Silent Mode';
+          } else if (this.setPositionMode === '0') {
+            this.Mode = 'Performance Mode';
+          } else {
+            this.Mode = 'Default Mode';
+          }
+          this.debugLog(`${this.accessory.displayName} Mode: ${this.Mode}`);
+          const adjustedMode = this.setPositionMode || 'ff';
           const payload = {
             commandType: 'command',
             command: 'setPosition',
-            parameter: `0,ff,${adjustedTargetPosition}`,
+            parameter: `0,${adjustedMode},${adjustedTargetPosition}`,
           } as payload;
 
           this.infoLog(
@@ -613,18 +691,21 @@ export class Curtain {
     } else {
       this.windowCoveringService.updateCharacteristic(this.platform.Characteristic.CurrentPosition, Number(this.CurrentPosition));
       this.debugLog(`Curtain: ${this.accessory.displayName} updateCharacteristic CurrentPosition: ${this.CurrentPosition}`);
+      this.mqttPublish('CurrentPosition', this.CurrentPosition);
     }
     if (this.PositionState === undefined) {
       this.debugLog(`Curtain: ${this.accessory.displayName} PositionState: ${this.PositionState}`);
     } else {
       this.windowCoveringService.updateCharacteristic(this.platform.Characteristic.PositionState, Number(this.PositionState));
       this.debugLog(`Curtain: ${this.accessory.displayName} updateCharacteristic PositionState: ${this.PositionState}`);
+      this.mqttPublish('PositionState', this.PositionState);
     }
     if (this.TargetPosition === undefined || Number.isNaN(this.TargetPosition)) {
       this.debugLog(`Curtain: ${this.accessory.displayName} TargetPosition: ${this.TargetPosition}`);
     } else {
       this.windowCoveringService.updateCharacteristic(this.platform.Characteristic.TargetPosition, Number(this.TargetPosition));
       this.debugLog(`Curtain: ${this.accessory.displayName} updateCharacteristic TargetPosition: ${this.TargetPosition}`);
+      this.mqttPublish('TargetPosition', this.TargetPosition);
     }
     if (!this.device.curtain?.hide_lightsensor) {
       if (this.CurrentAmbientLightLevel === undefined || Number.isNaN(this.CurrentAmbientLightLevel)) {
@@ -632,6 +713,7 @@ export class Curtain {
       } else {
         this.lightSensorService?.updateCharacteristic(this.platform.Characteristic.CurrentAmbientLightLevel, this.CurrentAmbientLightLevel);
         this.debugLog(`Curtain: ${this.accessory.displayName}` + ` updateCharacteristic CurrentAmbientLightLevel: ${this.CurrentAmbientLightLevel}`);
+	this.mqttPublish('CurrentAmbientLightLevel', this.CurrentAmbientLightLevel);
       }
     }
     if (this.device.ble) {
@@ -640,12 +722,14 @@ export class Curtain {
       } else {
         this.batteryService?.updateCharacteristic(this.platform.Characteristic.BatteryLevel, this.BatteryLevel);
         this.debugLog(`Curtain: ${this.accessory.displayName} updateCharacteristic BatteryLevel: ${this.BatteryLevel}`);
+	this.mqttPublish('BatteryLevel', this.BatteryLevel);
       }
       if (this.StatusLowBattery === undefined) {
         this.debugLog(`Curtain: ${this.accessory.displayName} StatusLowBattery: ${this.StatusLowBattery}`);
       } else {
         this.batteryService?.updateCharacteristic(this.platform.Characteristic.StatusLowBattery, this.StatusLowBattery);
         this.debugLog(`Curtain: ${this.accessory.displayName} updateCharacteristic StatusLowBattery: ${this.StatusLowBattery}`);
+	this.mqttPublish('StatusLowBattery', this.StatusLowBattery);
       }
     }
   }
@@ -702,6 +786,7 @@ export class Curtain {
     this.debugLog(`Curtain: ${this.accessory.displayName} TargetPosition: ${value}`);
 
     this.TargetPosition = value;
+    this.mqttPublish('TargetPosition', this.TargetPosition);
 
     await this.setMinMax();
     if (value > this.CurrentPosition) {
@@ -827,6 +912,21 @@ export class Curtain {
       this.deviceLogging = this.accessory.context.logging = 'standard';
       this.debugLog(`Curtain: ${this.accessory.displayName} Logging Not Set, Using: ${this.deviceLogging}`);
     }
+  }
+
+  FirmwareRevision(accessory: PlatformAccessory<Context>, device: device & devicesConfig): CharacteristicValue {
+    let FirmwareRevision: string;
+    this.debugLog(`Color Bulb: ${this.accessory.displayName} accessory.context.FirmwareRevision: ${accessory.context.FirmwareRevision}`);
+    this.debugLog(`Color Bulb: ${this.accessory.displayName} device.firmware: ${device.firmware}`);
+    this.debugLog(`Color Bulb: ${this.accessory.displayName} this.platform.version: ${this.platform.version}`);
+    if (accessory.context.FirmwareRevision) {
+      FirmwareRevision = accessory.context.FirmwareRevision;
+    } else if (device.firmware) {
+      FirmwareRevision = device.firmware;
+    } else {
+      FirmwareRevision = this.platform.version;
+    }
+    return FirmwareRevision;
   }
 
   minStep(device: device & devicesConfig): number {